test: split the two cross-cutting test files along module lines
test_adapter_args.py covered five modules and test_evaluators.py three; both cut
across src/ only because nobody split them — test_evaluators.py already carried
three '# ---' separators marking the seams.

The sorted set of test-function names is identical before and after, and coverage
stays at 100%, so no test was dropped or weakened. The only duplication introduced
is the four-line _ctx helper in the three evaluator files, which the shared-helpers
commit removes next.
